Base work of project replication

+++ server-plugin/server-plugin-report-coverage/src/main/java/io/onedev/server/plugin/report/coverage/PublishCoverageReportStep.java
@@ -6,15 +6,14 @@ import java.io.IOException;
 import java.io.OutputStream;
 import java.io.Serializable;
 import java.util.Map;
-import java.util.concurrent.Callable;
 
 import javax.annotation.Nullable;
 
+import io.onedev.server.entitymanager.ProjectManager;
 import org.apache.commons.lang.SerializationUtils;
 
 import io.onedev.commons.utils.ExceptionUtils;
 import io.onedev.commons.utils.FileUtils;
-import io.onedev.commons.utils.LockUtils;
 import io.onedev.commons.utils.TaskLogger;
 import io.onedev.server.OneDev;
 import io.onedev.server.buildspec.step.PublishReportStep;
@@ -24,6 +23,9 @@ import io.onedev.server.model.CoverageMetric;
 import io.onedev.server.persistence.dao.Dao;
 import io.onedev.server.annotation.Editable;
 
+import static io.onedev.commons.utils.LockUtils.write;
+import static io.onedev.server.plugin.report.coverage.CoverageReport.getReportLockName;
+
 @Editable
 public abstract class PublishCoverageReportStep extends PublishReportStep {
 
@@ -31,28 +33,25 @@ public abstract class PublishCoverageReportStep extends PublishReportStep {
 	
 	@Override
 	public Map<String, byte[]> run(Build build, File inputDir, TaskLogger logger) {
-		CoverageReport report = LockUtils.write(CoverageReport.getReportLockName(build), new Callable<CoverageReport>() {
+		CoverageReport report = write(getReportLockName(build), () -> {
+			File reportDir = new File(build.getStorageDir(), CoverageReport.DIR_CATEGORY + "/" + getReportName());
 
-			@Override
-			public CoverageReport call() throws Exception {
-				File reportDir = new File(build.getDir(), CoverageReport.CATEGORY + "/" + getReportName());
-
-				FileUtils.createDir(reportDir);
-				try {
-					CoverageReport report = createReport(build, inputDir, reportDir, logger);
-					if (report != null) {
-						report.writeTo(reportDir);
-						return report;
-					} else {
-						FileUtils.deleteDir(reportDir);
-						return null;
-					}
-				} catch (Exception e) {
+			FileUtils.createDir(reportDir);
+			try {
+				CoverageReport aReport = createReport(build, inputDir, reportDir, logger);
+				if (aReport != null) {
+					aReport.writeTo(reportDir);
+					OneDev.getInstance(ProjectManager.class).directoryModified(
+							build.getProject().getId(), reportDir.getParentFile());
+					return aReport;
+				} else {
 					FileUtils.deleteDir(reportDir);
-					throw ExceptionUtils.unchecked(e);
+					return null;
 				}
+			} catch (Exception e) {
+				FileUtils.deleteDir(reportDir);
+				throw ExceptionUtils.unchecked(e);
 			}
-			
 		});
 		
 		if (report != null) {
@@ -80,8 +79,8 @@ public abstract class PublishCoverageReportStep extends PublishReportStep {
 	protected abstract CoverageReport createReport(Build build, File inputDir, File reportDir, TaskLogger logger);
 
 	protected void writeLineCoverages(Build build, String blobPath, Map<Integer, CoverageStatus> lineCoverages) {
-		File reportDir = new File(build.getDir(), CoverageReport.CATEGORY + "/" + getReportName());
-		File lineCoverageFile = new File(reportDir, CoverageReport.FILES_DIR + "/" + blobPath);
+		File reportDir = new File(build.getStorageDir(), CoverageReport.DIR_CATEGORY + "/" + getReportName());
+		File lineCoverageFile = new File(reportDir, CoverageReport.DIR_FILES + "/" + blobPath);
 		FileUtils.createDir(lineCoverageFile.getParentFile());
 		try (OutputStream os = new FileOutputStream(lineCoverageFile)) {
 			SerializationUtils.serialize((Serializable) lineCoverages, os);
